renamed mod `systems` to `serial_port`
