+++ src/instructions/sync_write.rs
@@ -1,6 +1,6 @@
 use super::{instruction_id, packet_id, SyncWriteData};
 use crate::endian::{write_u16_le, write_u32_le};
-use crate::systems::SerialPort;
+use crate::serial_port::SerialPort;
 use crate::{Bus, WriteError};
 
 impl<ReadBuffer, WriteBuffer, T> Bus<ReadBuffer, WriteBuffer, T>
