+++ src/instructions/write.rs
@@ -1,6 +1,6 @@
 use super::{instruction_id, read_response_if_not_broadcast};
 use crate::endian::{write_u16_le, write_u32_le};
-use crate::systems::SerialPort;
+use crate::serial_port::SerialPort;
 use crate::{Bus, Response, TransferError};
 
 impl<ReadBuffer, WriteBuffer, T> Bus<ReadBuffer, WriteBuffer, T>
